+++ test/SelfOrganizingMapTest/add_binary_section.cpp
@@ -26,7 +26,7 @@ void add_binary_section(std::stringstream& ss, std::vector<std::vector<float>> c
     ss.write(reinterpret_cast<const char*>(&width), sizeof(int));
     ss.write(reinterpret_cast<const char*>(&height), sizeof(int));
     ss.write(reinterpret_cast<const char*>(&images[0][0]),
-        static_cast<std::streamsize>(width * height * sizeof(float)));
+        static_cast<std::streamsize>(static_cast<size_t>(width * height) * sizeof(float)));
     ss.write(reinterpret_cast<const char*>(&images[1][0]),
-        static_cast<std::streamsize>(width * height * sizeof(float)));
+        static_cast<std::streamsize>(static_cast<size_t>(width * height) * sizeof(float)));
 }
